+++ installer/conf/container.conf
@@ -9,29 +9,10 @@
 
 # Container inventory
 <source>
-	type omi
-	run_interval 60s
-	tag oms.container.containerinventory
-	items [
-		["root/cimv2","Container_ContainerInventory"]
-	]
-</source>
-
-<source>
   type containerinventory
   tag oms.containerinsights.ContainerInventory
   run_interval 60s
   log_level debug
-</source>
-
-# Image inventory
-<source>
-	type omi
-	run_interval 60s
-	tag oms.container.imageinventory
-	items [
-		["root/cimv2","Container_ImageInventory"]
-	]
 </source>
 
 # Container host inventory
@@ -52,15 +33,6 @@
   log_level debug
 </source>
 
-# Filter for correct format to endpoint
-<filter oms.container.containerinventory oms.container.imageinventory>
-	type filter_container
-</filter>
-
-<match oms.containerinsights.ContainerInventory**>
-  @type stdout
-</match>
-
 <match oms.api.ContainerNodeInventory**>
   type out_oms_api
   log_level debug
@@ -74,33 +46,22 @@
   max_retry_wait 9m
 </match>
 
-<match oms.container.containerinventory**>	
+<match oms.containerinsights.ContainerInventory**>
   type out_oms
   log_level debug
+  num_threads 5
   buffer_chunk_limit 20m
   buffer_type file
-  buffer_path %STATE_DIR_WS%/out_oms_containerinventory*.buffer
+  buffer_path /var/opt/microsoft/omsagent/62a84175-b18d-48ff-8b42-0694dda7b8bf/state/state/out_oms_containerinventory*.buffer
   buffer_queue_limit 20
+  buffer_queue_full_action drop_oldest_chunk
   flush_interval 20s
   retry_limit 10
-  retry_wait 15s
+  retry_wait 30s
   max_retry_wait 9m
 </match>
 
-<match oms.container.imageinventory**>
-  type out_oms
-  log_level debug
-  buffer_chunk_limit 20m
-  buffer_type file
-  buffer_path %STATE_DIR_WS%/out_oms_imageinventory*.buffer
-  buffer_queue_limit 20
-  flush_interval 20s
-  retry_limit 10
-  retry_wait 15s
-  max_retry_wait 9m
-</match>
-
-<match oms.api.cadvisorperf**>	
+<match oms.api.cadvisorperf**>
   type out_oms
   log_level debug
   num_threads 5
